Update dotnet SDK version and modify API call intervals
The primary changes in this commit include the update of the dotnet SDK version from 8.0.204 to 8.0.205 in multiple files. This update implicates both PresentationBuildTasks and runtimeIdentifierGraphPath as well. Also, there are modifications in the MainWindow.xaml.cs file which adjust the intervals for API calls, switching them from an hour to ten minutes. These changes aim to increase the weather update frequency.